requirements.txt
Pillow==10.1.0
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0